python_ingestion/requirements.txt
pandas
snowflake-connector-python
pyarrow
python-dotenv